Change way of getting user id

See merge request epamupskill/hw/book-notes!4

+++ book_notes/src/main/java/com/epamupskills/book_notes/domain/models/Book.kt
@@ -7,5 +7,5 @@ data class Book(
     val description: String,
     val thumbnailUrl: String,
     val isBookmarked: Boolean = false,
-    val noteId: Int? = null,
+    val noteId: Long? = null,
 )
